Change chassis control.

+++ dev/control/chassis.h
@@ -49,7 +49,7 @@ typedef struct {
     int target_rpm;
     int16_t target_current;
     uint16_t actual_angle;
-    uint16_t actual_rpm;
+    int actual_rpm;
     int16_t actual_current;
     uint8_t actual_temperature;
 } motor_t;
